style: :truck: Moved event directory
