+++ src/app/components/shared/tag-form/tag-form.component.html
@@ -1,6 +1,6 @@
 
 <div class="bg-white h-screen bg-opacity-50">
-    <div class="w-screen fixed b-0 rounded-tl-xl bg-light-blue p-2">
+    <div class="w-screen fixed rounded-tl-xl bg-light-blue p-2">
         <div class="flex">
             <app-tag tagLabel="Food" tagName="Food"></app-tag>
             <app-tag tagLabel="Lodging" tagName="Lodging"></app-tag>
